requirements.txt
asgiref==3.2.10
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
Django==3.1.3
django-cors-headers==3.5.0
django-model-utils==4.0.0
djangorestframework==3.12.2
idna==2.10
lxml==4.5.2
mysqlclient==2.0.1
numpy==1.19.4
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.26.1
webdriver-manager==3.2.2
